Send notification test + clean up.
Clean up info: Apparently some of the controllers were being called twice. This was because when app.js routed to a page, we already included the controller to be passed along with it. However, in the HTML files of the pages we were routing to, we included ng-controller in there. This made it so that our controllers were being called twice. I removed the ng-controller calls in the HTML files.